Claude Code: Read tracker task `97049fb9-8037-4a51-a82b-b33a5d667646` (Audio scene...
Prompt: Read tracker task `97049fb9-8037-4a51-a82b-b33a5d667646` (Audio scene redesign — Phase 3 implementation) via `tracker_get_task` and implement the **current** iteration described in the description: the math-function emission fix (Floor.of() integer-identity + cast-fallback, and sibling math function emission classes).

The investigation memo at `docs/plans/audio-scene-redesign/METAL_AMBIGUITY_BATCHED_RENDERER.md` is **mandatory reading first** — it contains the captured Metal source, the captured compiler error, the root cause analysis, and the specific Expression API locations to use. The fix specification in deliverable 1 of the current iteration is grounded in that memo's recommendation.

Verification on Mac/Metal is mandatory. Capture a verbatim excerpt of the new (working) Metal source after the fix as evidence that `floor((long)global_id)` no longer appears in the generated code.

The previous iterations are already complete — do not redo any of them. The job is structured as: read memo, fix Floor.of(), audit and fix siblings, verify on Mac, append implementation update to memo.
Session: 5e6f0cee-3239-4b83-9177-ed91a6fc9d57
Exit code: 0

+++ base/code/src/main/java/io/almostrealism/expression/Exponent.java
@@ -108,12 +108,16 @@ public class Exponent extends Expression<Double> {
 	/**
 	 * Creates and post-processes a power expression for the given base and exponent.
 	 *
+	 * <p>Non-FP operands are widened with {@link Expression#toDouble()} so the emitted
+	 * {@code pow()} call always has floating-point arguments — Metal's {@code pow()}
+	 * overloads are FP-only and an integer operand triggers a compile ambiguity.</p>
+	 *
 	 * @param base     the base expression
 	 * @param exponent the exponent expression
 	 * @return a simplified or constant-folded expression
 	 */
 	public static Expression<Double> of(Expression<Double> base, Expression<Double> exponent) {
-		return Expression.process(create(base, exponent));
+		return Expression.process(create(base.toDouble(), exponent.toDouble()));
 	}
 
 	/**
